NEURO_RAG_BACKEND/requirements.txt
langchain-openai==0.2.6
langchain
langchain_core
langgraph
langgraph-checkpoint
langgraph-checkpoint-sqlite
langchain-community
langgraph-supervisor
langdetect
agno

psycopg2-binary
tabulate
sqlalchemy
teradatasql
sqlalchemy-teradata

azure-cognitiveservices-speech
azure-cognitiveservices-search-customsearch
azure-identity
azure-search-documents==11.4.0
azure-keyvault-secrets
azure-functions
azurefunctions-extensions-bindings-blob
azure-cosmos==4.5.1
azure-mgmt-web
azure-mgmt-storage
azure-storage-blob
tiktoken
semantic-kernel

llama-index
llama-index-llms-azure-openai
llama-index-embeddings-azure-openai
pyodbc
azure.functions
sqlparse

fastapi
flask
openai
uvicorn
pytz
requests
python-dotenv
beautifulsoup4
#pyautogen==0.3.0
aiohttp==3.10.5
fuzzywuzzy
pandas
build
click
colorama
pytest

python-Levenshtein
pydantic==2.9.2

pyhive
impala